Refactor Metadata matching logic into separate module
